move test files into src
