+++ client/src/__tests__/App.test.tsx
@@ -1,5 +1,5 @@
 import { render, screen } from "@testing-library/react";
-import App from "../src/App";
+import App from "../App";
 
 test("renders a login prompt", () => {
 	render(<App />);
