+++ client/tsconfig.json
@@ -16,5 +16,5 @@
 		"noEmit": true,
 		"jsx": "react-jsx"
 	},
-	"include": ["src", "__tests__"]
+	"include": ["src"]
 }
